Alteration to Move class
main.py - moved move instantiation to above pokemon so that it can be
passed into the Move property as opposed to the "Punch" string.
- added print of pokemon.Move.Name as a proof that now our pokemon
object has a Move object encapsulated inside it

move.py - renamed Moves to Move just for now, once it is an array of
Moves it can be switched back.
move.pyc - need to ask clairethebear what the fuck this yoke is.