Introduction of Python style Properties
assets.py - Changed GameModel name field to be a Python style Property
after a little research. Decided on this as it is the neatest and most
literal, and provides ONE way of accessing name, unlike using a Java
style Getter and Setter

main.py - changed references to reflect the above change.